Finally, a python/c# back and forth call.
Notably, I'm interacting with a lot of generics on the C# side. That might be a bit of a nuisance in the long run. I've got to cut the amount of code by 50% to make this a clean implementation.

On the python side, I assume that all C# clients have the same proto objects (seems reasonable), and the first contact is to get the mapping info for those classes.

On the C# side, I make an assumption about what a handler looks like (it promises a ProtoMessage, and takes in a Protobuf object as its only paramters). Then it will register those objects at enable/disable to handle any incoming objects.

The idea is to minimize the amount of C# coding that needs to happen to control the domains (assuming that the primary audience is research and not game dev). Hence relying on generics and hiding all this from user .