Screenshot helper added and works all fine and dandy (converts main camera directly to byte array and matches requested size).
Woohoo, the render part of the render farm is coming together. Quite exciting, next piece will be sending objects to be rendered.

There is a lot of complicated layering going on right now. I'd like to build the V0 with the vines, then sit back and assess what exactly happened.

Once v0 of vines, the next demo would be RL (send an action, act and get a response stepwise). Then we could build the hardmaze in unity as well. Not getting ahead of myself...